server/requirements.txt
fastapi>=0.115
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.9
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openpyxl
apscheduler
